pyproject.toml
[project]
name = "americasnlp-2026"
version = "0.1.0"
description = "LLM-assisted RBMT for AmericasNLP 2026 image captioning"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "yaduha",
    "yaduha-hch",
    "yaduha-ovp",
    "yaduha-bzd",
    "yaduha-grn",
    "yaduha-yua",
    "yaduha-nlv",
    "sacrebleu>=2.0",
    "anthropic>=0.50",
    "openai>=1.40",  # transitively required by yaduha (imports openai types)
    "pydantic>=2.0",
    "Pillow>=10.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
# COMET pulls torch + a 1+ GB checkpoint. Opt in only when you need it:
#     uv sync --extra comet
comet = [
    "unbabel-comet>=2.2",
]

[project.scripts]
americasnlp = "americasnlp.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/americasnlp"]

[tool.uv.workspace]
members = [
    "yaduha",
    "yaduha-hch",
    "yaduha-ovp",
    "yaduha-bzd",
    "yaduha-grn",
    "yaduha-yua",
    "yaduha-nlv",
]

[tool.uv.sources]
yaduha     = { workspace = true, editable = true }
yaduha-hch = { workspace = true, editable = true }
yaduha-ovp = { workspace = true, editable = true }
yaduha-bzd = { workspace = true, editable = true }
yaduha-grn = { workspace = true, editable = true }
yaduha-yua = { workspace = true, editable = true }
yaduha-nlv = { workspace = true, editable = true }
